updating seeders

+++ backend/seeders/20250218143247-demo-personal-records.js
@@ -4,14 +4,18 @@ module.exports = {
   up: async () => {
     await PersonalRecord.bulkCreate([
       {
-        user_id: 1,
+        user_id: 1001,
+        session_id: 1001,
+        session_activity_id: 1001,
         activity_type: 'run',
         distance: 5000,
         best_time: 1200, 
         record_date: new Date('2025-01-20'),
       },
       {
-        user_id: 2,
+        user_id: 1002,
+        session_id: 1002,
+        session_activity_id: 1002,
         activity_type: 'bike',
         distance: 50000, 
         best_time: 5400, 
@@ -21,6 +25,6 @@ module.exports = {
   },
 
   down: async () => {
-    await PersonalRecord.destroy({ where: {} });
+    await PersonalRecord.destroy({ where: { user_id: [4, 5] } });
   }
 };
